ratemyreads/requirements.txt
django==4.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
drf-spectacular==0.27.2
